requirements.txt
Flask>=2.0.0
Flask-CORS>=3.0.0
pandas>=1.5.0
qrcode[pil]>=7.0.0
cryptography>=3.0.0
Pillow>=8.0.0
openpyxl>=3.0.0
python-dotenv>=0.19.0
pytz>=2021.1
gunicorn>=20.1.0
requests>=2.25.0
mailtrap>=3.0.0
